+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 #### Fixed
 - Fixed bug in `dallinger constraints generate` when using `pyproject.toml` in combination with a Git Dallinger dependency.
 
+#### Documentation
+- Made minor clarifications about the use of SSH keys and AWS permissions.
+
 ## [v12.0.0](https://github.com/dallinger/dallinger/tree/v12.0.0) (2025-12-16)
 
 #### Breaking Changes