+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 
 #### Fixed
+- Fixed bug where `server_pem` was not propagated to the Docker remote build process.
 - Pinned paramiko to <4.0.0 to address incompatibility with sshtunnel
   (see [discussion](https://github.com/pahaz/sshtunnel/issues/302)).
 - Fixed bug in `dallinger constraints generate` when using `pyproject.toml` in combination with a Git Dallinger dependency.