+++ CHANGELOG.md
@@ -1,6 +1,4 @@
-# Change Log
-
-## [Unreleased]
+## [v12.1.0](https://github.com/dallinger/dallinger/tree/v12.1.0) (2026-01-09)
 
 #### Changed
 - **docker-ssh server storage location moved to `~/.dallinger/docker-ssh/hosts`**: